Rename Python_Mini_Project.ipynb to Playlist_Recommendations.ipynb
This project involves creating a playlist recommendation system using Python. It analyzes user song preferences by comparing new user playlists with existing playlists in the dataset. If a match is found, the system recommends new songs that are not already in the user's playlist. The project utilizes Python data structures, loops, and conditional logic to find matching playlists and suggest relevant song recommendations.